+++ daemon/services/collectors/vm.go
@@ -8,10 +8,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/domalab/unraid-management-agent/daemon/domain"
-	"github.com/domalab/unraid-management-agent/daemon/dto"
-	"github.com/domalab/unraid-management-agent/daemon/lib"
-	"github.com/domalab/unraid-management-agent/daemon/logger"
+	"github.com/ruaan-deysel/unraid-management-agent/daemon/domain"
+	"github.com/ruaan-deysel/unraid-management-agent/daemon/dto"
+	"github.com/ruaan-deysel/unraid-management-agent/daemon/lib"
+	"github.com/ruaan-deysel/unraid-management-agent/daemon/logger"
 )
 
 type VMCollector struct {